+++ app/src/main/java/com/konkuk/medicarecall/data/api/elders/HealthService.kt
@@ -1,4 +1,4 @@
-package com.konkuk.medicarecall.data.api
+package com.konkuk.medicarecall.data.api.elders
 
 import com.konkuk.medicarecall.data.dto.response.HealthResponseDto
 import retrofit2.http.GET
