+++ app/src/main/java/com/konkuk/medicarecall/data/di/ApiModule.kt
@@ -1,23 +1,23 @@
 package com.konkuk.medicarecall.data.di
 
-import com.konkuk.medicarecall.data.api.ElderRegisterService
-import com.konkuk.medicarecall.data.api.EldersInfoService
-import com.konkuk.medicarecall.data.api.MemberRegisterService
-import com.konkuk.medicarecall.data.api.NaverPayService
+import com.konkuk.medicarecall.data.api.elders.ElderRegisterService
+import com.konkuk.medicarecall.data.api.elders.EldersInfoService
+import com.konkuk.medicarecall.data.api.member.MemberRegisterService
+import com.konkuk.medicarecall.data.api.payments.NaverPayService
 import com.konkuk.medicarecall.data.api.NoticeService
-import com.konkuk.medicarecall.data.api.SetCallService
+import com.konkuk.medicarecall.data.api.elders.SetCallService
 import com.konkuk.medicarecall.data.api.SettingService
-import com.konkuk.medicarecall.data.api.SubscribeService
+import com.konkuk.medicarecall.data.api.elders.SubscribeService
 import com.konkuk.medicarecall.data.api.TokenRefreshService
 import com.konkuk.medicarecall.data.api.VerificationService
 import com.konkuk.medicarecall.data.api.HomeService
-import com.konkuk.medicarecall.data.api.GlucoseService
-import com.konkuk.medicarecall.data.api.MealService
-import com.konkuk.medicarecall.data.api.MedicineService
-import com.konkuk.medicarecall.data.api.SleepService
-import com.konkuk.medicarecall.data.api.HealthService
-import com.konkuk.medicarecall.data.api.MentalService
-import com.konkuk.medicarecall.data.api.StatisticsService
+import com.konkuk.medicarecall.data.api.elders.GlucoseService
+import com.konkuk.medicarecall.data.api.elders.MealService
+import com.konkuk.medicarecall.data.api.elders.MedicineService
+import com.konkuk.medicarecall.data.api.elders.SleepService
+import com.konkuk.medicarecall.data.api.elders.HealthService
+import com.konkuk.medicarecall.data.api.elders.MentalService
+import com.konkuk.medicarecall.data.api.elders.StatisticsService
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
