+++ app/src/main/java/com/konkuk/medicarecall/data/repositoryimpl/GlucoseRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.konkuk.medicarecall.data.repositoryimpl
 
-import com.konkuk.medicarecall.data.api.GlucoseService
+import com.konkuk.medicarecall.data.api.elders.GlucoseService
 import com.konkuk.medicarecall.data.repository.GlucoseRepository
 import com.konkuk.medicarecall.data.dto.response.GlucoseResponseDto
 import retrofit2.HttpException
