+++ app/src/main/java/com/konkuk/medicarecall/data/repositoryimpl/MedicineRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.konkuk.medicarecall.data.repositoryimpl
 
-import com.konkuk.medicarecall.data.api.MedicineService
+import com.konkuk.medicarecall.data.api.elders.MedicineService
 import com.konkuk.medicarecall.data.repository.EldersHealthInfoRepository
 import com.konkuk.medicarecall.data.repository.MedicineRepository
 import com.konkuk.medicarecall.ui.feature.homedetail.medicine.viewmodel.DoseStatus
