+++ app/src/main/java/com/konkuk/medicarecall/data/repositoryimpl/SleepRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.konkuk.medicarecall.data.repositoryimpl
 
-import com.konkuk.medicarecall.data.api.SleepService
+import com.konkuk.medicarecall.data.api.elders.SleepService
 import com.konkuk.medicarecall.data.repository.SleepRepository
 import com.konkuk.medicarecall.ui.feature.homedetail.sleep.viewmodel.SleepUiState
 import java.time.LocalDate
